Move JournalFileStorage-related classes to journal/_file.py
