+++ optuna/storages/__init__.py
@@ -7,10 +7,10 @@ from optuna.storages._cached_storage import _CachedStorage
 from optuna.storages._heartbeat import fail_stale_trials
 from optuna.storages._in_memory import InMemoryStorage
 from optuna.storages._rdb.storage import RDBStorage
-from optuna.storages.journal._backend import JournalFileStorage
 from optuna.storages.journal._base import BaseJournalLogStorage
-from optuna.storages.journal._file_lock import JournalFileOpenLock as _JournalFileOpenLock
-from optuna.storages.journal._file_lock import JournalFileSymlinkLock as _JournalFileSymlinkLock
+from optuna.storages.journal._file import JournalFileOpenLock as _JournalFileOpenLock
+from optuna.storages.journal._file import JournalFileStorage
+from optuna.storages.journal._file import JournalFileSymlinkLock as _JournalFileSymlinkLock
 from optuna.storages.journal._redis import JournalRedisStorage
 from optuna.storages.journal._storage import JournalStorage
