+++ optuna/storages/journal/__init__.py
@@ -1,7 +1,7 @@
-from optuna.storages.journal._backend import JournalFileBackend
 from optuna.storages.journal._base import BaseJournalBackend
-from optuna.storages.journal._file_lock import JournalFileOpenLock
-from optuna.storages.journal._file_lock import JournalFileSymlinkLock
+from optuna.storages.journal._file import JournalFileBackend
+from optuna.storages.journal._file import JournalFileOpenLock
+from optuna.storages.journal._file import JournalFileSymlinkLock
 from optuna.storages.journal._redis import JournalRedisBackend
 from optuna.storages.journal._storage import JournalStorage
